requirements.txt
Flask==3.0.0
line-bot-sdk==3.5.0
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
